portal/requirements.txt
# PowerUp Portal — Python dependencies
# Install with: pip install -r requirements.txt

# ── Web framework ─────────────────────────────────────────────
streamlit>=1.35.0

# ── Google APIs ───────────────────────────────────────────────
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.126.0

# ── Environment / config ──────────────────────────────────────
python-dotenv>=1.0.1

# ── Data handling ─────────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2        # Excel read/write (M3 client upload, migration)

# ── HTTP ──────────────────────────────────────────────────────
requests>=2.31.0       # M1 Apps Script HTTP trigger

# ── M2 / M3 automation dependencies (kept until full Slides migration) ──
matplotlib>=3.8.0      # M2 chart generation
lxml>=5.2.0            # M2 PPTX XML manipulation
python-pptx==0.6.23    # M2 + M3 deck generation (until migrated to Slides API)
